refactor: privacy page
